Rename ErrorDetail --> ErrorMessage::Detail
More accurately reflects its intention and groups
it with its co-dependendants.
