+++ .rubocop_todo.yml
@@ -973,7 +973,7 @@ RSpec/DescribedClass:
     - 'spec/presenters/claim_history_presenter_spec.rb'
     - 'spec/presenters/determination_presenter_spec.rb'
     - 'spec/presenters/error_detail_collection_spec.rb'
-    - 'spec/presenters/error_detail_spec.rb'
+    - 'spec/presenters/error_message/detail_spec.rb'
     - 'spec/presenters/error_message_translator_spec.rb'
     - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/presenters/expense_presenter_spec.rb'
@@ -1924,7 +1924,7 @@ RSpec/MultipleExpectations:
     - 'spec/presenters/claim/base_claim_presenter_spec.rb'
     - 'spec/presenters/claim_state_transition_presenter_spec.rb'
     - 'spec/presenters/determination_presenter_spec.rb'
-    - 'spec/presenters/error_detail_spec.rb'
+    - 'spec/presenters/error_message/detail_spec.rb'
     - 'spec/presenters/error_message_translator_spec.rb'
     - 'spec/presenters/fixed_fee_presenter_spec.rb'
     - 'spec/presenters/graduated_fee_presenter_spec.rb'
@@ -2050,7 +2050,7 @@ RSpec/MultipleMemoizedHelpers:
     - 'spec/models/timed_transitions/batch_transitioner_spec.rb'
     - 'spec/presenters/claim/base_claim_presenter_spec.rb'
     - 'spec/presenters/claim_history_presenter_spec.rb'
-    - 'spec/presenters/error_detail_spec.rb'
+    - 'spec/presenters/error_message/detail_spec.rb'
     - 'spec/presenters/management_information_presenter_spec.rb'
     - 'spec/requests/external_users/expenses/distance_calculation_spec.rb'
     - 'spec/services/claims/fee_calculator/calculate_price_spec.rb'
@@ -3399,7 +3399,7 @@ Style/NumericLiterals:
     - 'spec/models/expense_spec.rb'
     - 'spec/models/fee/base_fee_spec.rb'
     - 'spec/models/timed_transitions/batch_transitioner_spec.rb'
-    - 'spec/presenters/error_detail_spec.rb'
+    - 'spec/presenters/error_message/detail_spec.rb'
     - 'spec/presenters/error_message/presenter_spec.rb'
     - 'spec/presenters/expense_presenter_spec.rb'
     - 'spec/presenters/fee_presenter_spec.rb'
