+++ app/presenters/error_message/detail_collection.rb
@@ -1,5 +1,5 @@
-# This class holds a collection of ErrorDetail objects, keyed by fieldname to which error it pertains.
-# Each key can hold more than one ErrorDetail.  The class provides specialised methods for retrieving
+# This class holds a collection of ErrorMessage::Detail objects, keyed by fieldname to which error it pertains.
+# Each key can hold more than one ErrorMessage::Detail.  The class provides specialised methods for retrieving
 # short messages by fieldname, and all the long messages with associated fieldnames
 #
 module ErrorMessage
