+++ spec/presenters/error_message/presenter_spec.rb
@@ -35,9 +35,9 @@ RSpec.describe ErrorMessage::Presenter do
 
       it do
         is_expected.to eq(
-          [ErrorDetail.new(:date_of_birth,
-                           'The date of birth may not be more than 100 years old', 'Enter a valid date',
-                           'The date of birth is too early', 20)]
+          [ErrorMessage::Detail.new(:date_of_birth,
+                                    'The date of birth may not be more than 100 years old', 'Enter a valid date',
+                                    'The date of birth is too early', 20)]
         )
       end
     end
@@ -48,7 +48,7 @@ RSpec.describe ErrorMessage::Presenter do
       it do
         is_expected.to eq(
           [
-            ErrorDetail.new(:date_of_birth, 'Date of birth foo bar', 'Foo bar', 'Date of birth foo bar')
+            ErrorMessage::Detail.new(:date_of_birth, 'Date of birth foo bar', 'Foo bar', 'Date of birth foo bar')
           ]
         )
       end
@@ -60,7 +60,7 @@ RSpec.describe ErrorMessage::Presenter do
       it do
         expect(presenter.header_errors).to eq(
           [
-            ErrorDetail.new(:defendant_2_name, 'Defendant 2 name is invalid', 'Is invalid', 'Defendant 2 name is invalid')
+            ErrorMessage::Detail.new(:defendant_2_name, 'Defendant 2 name is invalid', 'Is invalid', 'Defendant 2 name is invalid')
           ]
         )
       end
@@ -72,7 +72,7 @@ RSpec.describe ErrorMessage::Presenter do
       it do
         is_expected.to eq(
           [
-            ErrorDetail.new(:defendant_2_first_name, 'Enter a first name for the second defendant', 'Enter a first name', 'The first name for the second defendant must not be blank')
+            ErrorMessage::Detail.new(:defendant_2_first_name, 'Enter a first name for the second defendant', 'Enter a first name', 'The first name for the second defendant must not be blank')
           ]
         )
       end
@@ -86,8 +86,8 @@ RSpec.describe ErrorMessage::Presenter do
 
       it do
         is_expected.to eq(
-          [ErrorDetail.new(:name, 'The claimant name must not be blank, please enter a name', 'Enter a name', 'The claimant name must not be blank', 50),
-           ErrorDetail.new(:name, 'The name cannot be longer than 50 characters', 'Too long', 'The name cannot be longer than 50 characters', 50)]
+          [ErrorMessage::Detail.new(:name, 'The claimant name must not be blank, please enter a name', 'Enter a name', 'The claimant name must not be blank', 50),
+           ErrorMessage::Detail.new(:name, 'The name cannot be longer than 50 characters', 'Too long', 'The name cannot be longer than 50 characters', 50)]
         )
       end
     end
